requirements.txt
fastapi==0.110.0
uvicorn==0.28.0
google-generativeai==0.4.1
vtracer==0.6.3
Pillow==10.2.0
